# [XLA] Fix and generalize to row_vectorization few waves