--- a/tensorflow/compiler/xla/service/gpu/tests/launch_dimensions.hlo
+++ b/tensorflow/compiler/xla/service/gpu/tests/launch_dimensions.hlo
@@ -11,3 +11,17 @@ ENTRY main {
   a = f32[100, 20]{1,0} parameter(0)
   ROOT b = f32[100, 20]{1,0} tanh(a)
 }
+
+// -----
+
+// This test that we cap grid launch code when few_waves is enabled.
+
+// CHECK:       !2 = !{i32 0, i32 1280}
+// CHECK:       !3 = !{i32 0, i32 128}
+
+HloModule Test
+
+ENTRY main {
+  a = f32[10000, 10000]{1,0} parameter(0)
+  ROOT b = f32[10000, 10000]{1,0} tanh(a)
+}