@nouiz 2021-04-13
One test is failing. But I have an updated version of this PR that also enable few_waves in combination of that recently merged PR:
https://github.com/tensorflow/tensorflow/pull/48381

I think it will be better to merge the both changes at the same time. I'll update this PR in a few days when I have checked that we do not have speed regression with it.