@nouiz 2021-04-21
I have the code working locally. I started today some new speed test to make sure it doesn't give regression in other cases. When it is all fine here, I'll update it.